Internal Documentation

Documentation for IterativeSolvers.jl's internals.

Internal Documentation

Index

ConvergenceHistory Internals

KrylovSubspace Internals

Other Functions

Index

IterativeSolvers.RestartedHistory

IterativeSolvers.Adivtype

IterativeSolvers.Amultype

IterativeSolvers.appendunit!

IterativeSolvers.harmonicrestart!

IterativeSolvers.idfact

IterativeSolvers.init!

IterativeSolvers.initrand!

IterativeSolvers.initrand!

IterativeSolvers.isconverged

IterativeSolvers.lastvec

IterativeSolvers.nextiter!

IterativeSolvers.nextvec

IterativeSolvers.orthogonalize

IterativeSolvers.plot_collection

IterativeSolvers.plotable

IterativeSolvers.randx

IterativeSolvers.reserve!

IterativeSolvers.setconv

IterativeSolvers.showplot

IterativeSolvers.shrink!

IterativeSolvers.thickrestart!

IterativeSolvers.zerox

ConvergenceHistory Internals

Typealiases

IterativeSolvers.PlainHistory
IterativeSolvers.RestartedHistory

Functions

IterativeSolvers.nextiter!
IterativeSolvers.reserve!
IterativeSolvers.shrink!
IterativeSolvers.setmvps
IterativeSolvers.setmtvps
IterativeSolvers.setconv
IterativeSolvers.showplot

KrylovSubspace Internals

Functions

IterativeSolvers.lastvec — Function.

lastvec(K)

Get last vector computed in the Krylov subspace K.

Arguments

K::KrylovSubspace: Krylov subspace.

Output

::Vector: last vector.

source

IterativeSolvers.nextvec — Function.

nextvec(K)

Compute next vector in the Krylov subspace K.

Arguments

K::KrylovSubspace: Krylov subspace.

Output

::Vector: next vector.

source

IterativeSolvers.init! — Function.

init!(K, v)

Initialize the KrylovSubspace K with a specified nonunit vector.

Arguments

K::KrylovSubspace: Krylov subspace.

v::Vector: initial vector.

source

IterativeSolvers.initrand! — Method.

initrand!(K)

Initialize the Krylov subspace K with a random unit vector.

Arguments

K::KrylovSubspace: Krylov subspace.

source

IterativeSolvers.appendunit! — Function.

appendunit!(K, w)

Append normalize(w) vector to krylob subspace K.

Arguments

K::KrylovSubspace: Krylov subspace.

w::Vector: vector to append.

source

IterativeSolvers.orthogonalize — Function.

orthogonalize{T}(v, K, p)

Orthogonalize a vector v against the last p basis vectors defined by the Krylov subspace K.

Arguments

v::Vector: vector to orthogonalize.

K::KrylovSubspace: subspace to orthogonalize v against.

p::Int=K.order: last p remembered vectors.

Keywords

method::Symbol=:ModifiedGramSchmidt: orthogonalization method. Choose from:

:GramSchmidt: Gram Schmidt method.

:ModifiedGramSchmidt: Modified Gram Schmidt method.

:Householder: Householder method.

normalize::Bool=false: normalize vector.

Output

::Vector: orthogonalized vector.

::Vector: orthogonalization coefficients.

source

Other Functions

IterativeSolvers.idfact
IterativeSolvers.isconverged
IterativeSolvers.thickrestart!
IterativeSolvers.harmonicrestart!
IterativeSolvers.plot_collection
IterativeSolvers.plotable
IterativeSolvers.Adivtype
IterativeSolvers.Amultype
IterativeSolvers.randx
IterativeSolvers.zerox
IterativeSolvers.update!
IterativeSolvers.initrand!(::Vector)