Public Documentation

Documentation for IterativeSolvers.jl's public interface.

Public Documentation

Index

Types

ConvergenceHistory

KrylovSubspace

Linear Solvers

Jacobi

Gauss-Seidel

Successive over relaxation

Symmetric successive over relaxation

IDRS

LSMR

LSQR

Conjugate gradients

Chebyshev iteration

Generalized minimal residual method (with restarts)

Eigen Solvers

Power iteration

Inverse power iteration

Simple Lanczos

Golub-Kahan-Lanczos

Randomized

Condition number estimate

Extremal eigenvalue estimates

Norm estimate

Randomized singular value decomposition

Index

IterativeSolvers.ConvergenceHistory

IterativeSolvers.KrylovSubspace

IterativeSolvers.cg

IterativeSolvers.cg!

IterativeSolvers.chebyshev

IterativeSolvers.chebyshev!

IterativeSolvers.eiglancz

IterativeSolvers.gauss_seidel

IterativeSolvers.gauss_seidel!

IterativeSolvers.gmres

IterativeSolvers.gmres!

IterativeSolvers.idrs

IterativeSolvers.idrs!

IterativeSolvers.invpowm

IterativeSolvers.jacobi

IterativeSolvers.jacobi!

IterativeSolvers.lsmr

IterativeSolvers.lsmr!

IterativeSolvers.lsqr

IterativeSolvers.lsqr!

IterativeSolvers.powm

IterativeSolvers.rcond

IterativeSolvers.reig

IterativeSolvers.reigmax

IterativeSolvers.reigmin

IterativeSolvers.rnorm

IterativeSolvers.rnorms

IterativeSolvers.rsvd_fnkz

IterativeSolvers.rsvdfact

IterativeSolvers.sor

IterativeSolvers.sor!

IterativeSolvers.ssor

IterativeSolvers.ssor!

IterativeSolvers.svdl

Types

ConvergenceHistory

IterativeSolvers.ConvergenceHistory — Type.

Store general and in-depth information about an iterative method.

Fields

mvps::Int: number of matrix vector products.

mtvps::Int: number of transposed matrix-vector products

iters::Int: iterations taken by the method.

restart::T: restart relevant information.

T == Int: iterations per restart.

T == Void: methods without restarts.

isconverged::Bool: convergence of the method.

data::Dict{Symbol,Any}: Stores all the information stored during the method execution. It stores tolerances, residuals and other information, e.g. ritz values in svdl.

Constructors

ConvergenceHistory()
ConvergenceHistory(restart)

Create ConvergenceHistory with empty fields.

Arguments

restart: number of iterations per restart.

Plots

Supports plots using the Plots.jl package via a type recipe. Vectors are ploted as series and matrices as scatterplots.

Implements

Base: getindex, setindex!, push!

source

KrylovSubspace

IterativeSolvers.KrylovSubspace — Type.

Collection of information on the Krylov subspace generated over the course of an iterative Krylov solver.

Fields

A: linear operator that generating the subspace.

n::Int: dimension of problem.

order::Int: order of subspace (maximum size).

v::Vector{Vector}: Krylov vectors.

mvps::Int: count of matrix-vector products.

Constructors

KrylovSubspace(A, order)
KrylovSubspace(A, n, order)
KrylovSubspace(A, n, order, T)
KrylovSubspace(A, n, order, v)

Create new KrylovSubspace.

Arguments

T::Type: type of the elements inside the Krylov vectors (eltype(v)).

v::Vector{Vector}: initial Krylov subspace.

KrylovSubspace(K, order=size(A,2), v=[])

Reset given KrylovSubspace A.

Arguments

K::KrylovSubspace: Krylov subspace to reset.

Implements

Base: append!, eye, size.

source

Linear Solvers

Jacobi

IterativeSolvers.jacobi — Function.

jacobi(A, b)

Solve A*x=b with the Jacobi method.

ch is a ConvergenceHistory object. Otherwise it will only return x. If log is set to true is given, method will output a tuple x, ch. Where The plot attribute can only be used when log is set version.

Arguments

A: linear operator.

Keywords

tol::Real = size(A,2)^3*eps(): stopping tolerance.

maxiter::Integer = size(A,2)^2: maximum number of iterations.

verbose::Bool = false: verbose flag.

log::Bool = false: output an extra element of type ConvergenceHistory containing extra information of the method execution.

plot::Bool = false: plot data. (Only when log is set)

Output

if log is false

x: approximated solution.

if log is true

x: approximated solution.

ch: convergence history.

ConvergenceHistory keys

:tol => ::Real: stopping tolerance.

:resnom => ::Vector: residual norm at each iteration.

source

IterativeSolvers.jacobi! — Function.

jacobi!(x, A, b)

Overwrite x.

Solve A*x=b with the Jacobi method.

ch is a ConvergenceHistory object. Otherwise it will only return x. If log is set to true is given, method will output a tuple x, ch. Where The plot attribute can only be used when log is set version.

Arguments

x: initial guess, overwrite final estimation.

A: linear operator.

Keywords

tol::Real = size(A,2)^3*eps(): stopping tolerance.

maxiter::Integer = size(A,2)^2: maximum number of iterations.

verbose::Bool = false: verbose flag.

log::Bool = false: output an extra element of type ConvergenceHistory containing extra information of the method execution.

plot::Bool = false: plot data. (Only when log is set)

Output

if log is false

x: approximated solution.

if log is true

x: approximated solution.

ch: convergence history.

ConvergenceHistory keys

:tol => ::Real: stopping tolerance.

:resnom => ::Vector: residual norm at each iteration.

source

Gauss-Seidel

IterativeSolvers.gauss_seidel — Function.

gauss_seidel(A, b)

Solve A*x=b with the Gauss-Seidel method.

ch is a ConvergenceHistory object. Otherwise it will only return x. If log is set to true is given, method will output a tuple x, ch. Where The plot attribute can only be used when log is set version.

Arguments

A: linear operator.

Keywords

tol::Real = size(A,2)^3*eps(): stopping tolerance.

maxiter::Integer = size(A,2)^2: maximum number of iterations.

verbose::Bool = false: verbose flag.

log::Bool = false: output an extra element of type ConvergenceHistory containing extra information of the method execution.

plot::Bool = false: plot data. (Only when log is set)

Output

if log is false

x: approximated solution.

if log is true

x: approximated solution.

ch: convergence history.

ConvergenceHistory keys

:tol => ::Real: stopping tolerance.

:resnom => ::Vector: residual norm at each iteration.

source

IterativeSolvers.gauss_seidel! — Function.

gauss_seidel!(x, A, b)

Overwrite x.

Solve A*x=b with the Gauss-Seidel method.

ch is a ConvergenceHistory object. Otherwise it will only return x. If log is set to true is given, method will output a tuple x, ch. Where The plot attribute can only be used when log is set version.

Arguments

x: initial guess, overwrite final estimation.

A: linear operator.

Keywords

tol::Real = size(A,2)^3*eps(): stopping tolerance.

maxiter::Integer = size(A,2)^2: maximum number of iterations.

verbose::Bool = false: verbose flag.

log::Bool = false: output an extra element of type ConvergenceHistory containing extra information of the method execution.

plot::Bool = false: plot data. (Only when log is set)

Output

if log is false

x: approximated solution.

if log is true

x: approximated solution.

ch: convergence history.

ConvergenceHistory keys

:tol => ::Real: stopping tolerance.

:resnom => ::Vector: residual norm at each iteration.

source

Successive over relaxation

IterativeSolvers.sor — Function.

sor(A, b, ω)

Solve A*x=b with the successive overrelaxation method.

ch is a ConvergenceHistory object. Otherwise it will only return x. If log is set to true is given, method will output a tuple x, ch. Where The plot attribute can only be used when log is set version.

Arguments

shift::Number=0: shift to be applied to matrix A.

A: linear operator.

Keywords

tol::Real = size(A,2)^3*eps(): stopping tolerance.

maxiter::Integer = size(A,2)^2: maximum number of iterations.

verbose::Bool = false: verbose flag.

log::Bool = false: output an extra element of type ConvergenceHistory containing extra information of the method execution.

plot::Bool = false: plot data. (Only when log is set)

Output

if log is false

x: approximated solution.

if log is true

x: approximated solution.

ch: convergence history.

ConvergenceHistory keys

:tol => ::Real: stopping tolerance.

:resnom => ::Vector: residual norm at each iteration.

source

IterativeSolvers.sor! — Function.

sor!(x, A, b, ω)

Overwrite x.

Solve A*x=b with the successive overrelaxation method.

ch is a ConvergenceHistory object. Otherwise it will only return x. If log is set to true is given, method will output a tuple x, ch. Where The plot attribute can only be used when log is set version.

Arguments

x: initial guess, overwrite final estimation.

shift::Number=0: shift to be applied to matrix A.

A: linear operator.

Keywords

tol::Real = size(A,2)^3*eps(): stopping tolerance.

maxiter::Integer = size(A,2)^2: maximum number of iterations.

verbose::Bool = false: verbose flag.

log::Bool = false: output an extra element of type ConvergenceHistory containing extra information of the method execution.

plot::Bool = false: plot data. (Only when log is set)

Output

if log is false

x: approximated solution.

if log is true

x: approximated solution.

ch: convergence history.

ConvergenceHistory keys

:tol => ::Real: stopping tolerance.

:resnom => ::Vector: residual norm at each iteration.

source

Symmetric successive over relaxation

IterativeSolvers.ssor — Function.

ssor(A, b, ω)

Solve A*x=b with the symmetric successive overrelaxation method.

ch is a ConvergenceHistory object. Otherwise it will only return x. If log is set to true is given, method will output a tuple x, ch. Where The plot attribute can only be used when log is set version.

Arguments

shift::Number=0: shift to be applied to matrix A.

A: linear operator.

Keywords

tol::Real = size(A,2)^3*eps(): stopping tolerance.

maxiter::Integer = size(A,2)^2: maximum number of iterations.

verbose::Bool = false: verbose flag.

log::Bool = false: output an extra element of type ConvergenceHistory containing extra information of the method execution.

plot::Bool = false: plot data. (Only when log is set)

Output

if log is false

x: approximated solution.

if log is true

x: approximated solution.

ch: convergence history.

ConvergenceHistory keys

:tol => ::Real: stopping tolerance.

:resnom => ::Vector: residual norm at each iteration.

source

IterativeSolvers.ssor! — Function.

ssor!(x, A, b, ω)

Overwrite x.

Solve A*x=b with the symmetric successive overrelaxation method.

ch is a ConvergenceHistory object. Otherwise it will only return x. If log is set to true is given, method will output a tuple x, ch. Where The plot attribute can only be used when log is set version.

Arguments

x: initial guess, overwrite final estimation.

shift::Number=0: shift to be applied to matrix A.

A: linear operator.

Keywords

tol::Real = size(A,2)^3*eps(): stopping tolerance.

maxiter::Integer = size(A,2)^2: maximum number of iterations.

verbose::Bool = false: verbose flag.

log::Bool = false: output an extra element of type ConvergenceHistory containing extra information of the method execution.

plot::Bool = false: plot data. (Only when log is set)

Output

if log is false

x: approximated solution.

if log is true

x: approximated solution.

ch: convergence history.

ConvergenceHistory keys

:tol => ::Real: stopping tolerance.

:resnom => ::Vector: residual norm at each iteration.

source

IDRS

IterativeSolvers.idrs — Function.

idrs(A, b)

Solve A*x=b using the induced dimension reduction method.

If log is set to true is given, method will output a tuple x, ch. Where ch is a ConvergenceHistory object. Otherwise it will only return x.

The plot attribute can only be used when log is set version.

Arguments

A: linear operator.

b: right hand side.

Keywords

Pl = 1: left preconditioner of the method.

Pr = 1: left preconditioner of the method.

tol::Real = sqrt(eps()): stopping tolerance.

restart::Integer = min(20,length(b)): maximum number of iterations per restart.

maxiter::Integer = min(20,length(b)): maximum number of iterations.

verbose::Bool = false: print method information.

log::Bool = false: output an extra element of type ConvergenceHistory containing extra information of the method execution.

plot::Bool = false: plot data. (Only when log is set)

Output

if log is false

x: approximated solution.

if log is true

x: approximated solution.

ch: convergence history.

ConvergenceHistory keys

:tol => ::Real: stopping tolerance.

:resnom => ::Vector: residual norm at each iteration.

source

IterativeSolvers.idrs! — Function.

idrs!(x, A, b)

Overwrite x.

Solve A*x=b using the induced dimension reduction method.

If log is set to true is given, method will output a tuple x, ch. Where ch is a ConvergenceHistory object. Otherwise it will only return x.

The plot attribute can only be used when log is set version.

Arguments

x: initial guess, overwrite final estimation.

A: linear operator.

b: right hand side.

Keywords

Pl = 1: left preconditioner of the method.

Pr = 1: left preconditioner of the method.

tol::Real = sqrt(eps()): stopping tolerance.

restart::Integer = min(20,length(b)): maximum number of iterations per restart.

maxiter::Integer = min(20,length(b)): maximum number of iterations.

verbose::Bool = false: print method information.

log::Bool = false: output an extra element of type ConvergenceHistory containing extra information of the method execution.

plot::Bool = false: plot data. (Only when log is set)

Output

if log is false

x: approximated solution.

if log is true

x: approximated solution.

ch: convergence history.

ConvergenceHistory keys

:tol => ::Real: stopping tolerance.

:resnom => ::Vector: residual norm at each iteration.

source

References

[1] IDR(s): a family of simple and fast algorithms for solving large
    nonsymmetric linear systems. P. Sonneveld and M. B. van Gijzen
    SIAM J. Sci. Comput. Vol. 31, No. 2, pp. 1035--1062, 2008
[2] Algorithm 913: An Elegant IDR(s) Variant that Efficiently Exploits
    Bi-orthogonality Properties. M. B. van Gijzen and P. Sonneveld
    ACM Trans. Math. Software,, Vol. 38, No. 1, pp. 5:1-5:19, 2011
[3] This file is a translation of the following MATLAB implementation:
    http://ta.twi.tudelft.nl/nw/users/gijzen/idrs.m
[4] IDR(s)' webpage http://ta.twi.tudelft.nl/nw/users/gijzen/IDR.html

LSMR

IterativeSolvers.lsmr — Function.

lsmr(A, b)

Minimize ||Ax-b||^2 + λ^2 ||x||^2 for A*x=b.

The method is based on the Golub-Kahan bidiagonalization process. It is algebraically equivalent to applying MINRES to the normal equation (ATA+λ2I)x=ATb, but has better numerical properties, especially if A is ill-conditioned.

If log is set to true is given, method will output a tuple x, ch. Where ch is a ConvergenceHistory object. Otherwise it will only return x.

The plot attribute can only be used when log is set version.

Arguments

A: linear operator.

b: right hand side.

Keywords

λ::Number = 0: lambda.

atol::Number = 1e-6, btol::Number = 1e-6: stopping tolerances. If both are 1.0e-9 (say), the final residual norm should be accurate to about 9 digits. (The final x will usually have fewer correct digits, depending on cond(A) and the size of damp).

conlim::Number = 1e8: stopping tolerance. lsmr terminates if an estimate of cond(A) exceeds conlim. For compatible systems Ax = b, conlim could be as large as 1.0e+12 (say). For least-squares problems, conlim should be less than 1.0e+8. Maximum precision can be obtained by setting atol = btol = conlim = zero, but the number of iterations may then be excessive.

maxiter::Integer = min(20,length(b)): maximum number of iterations.

verbose::Bool = false: print method information.

log::Bool = false: output an extra element of type ConvergenceHistory containing extra information of the method execution.

plot::Bool = false: plot data. (Only when log is set)

Output

if log is false

x: approximated solution.

if log is true

x: approximated solution.

ch: convergence history.

ConvergenceHistory keys

:atol => ::Real: atol stopping tolerance.

:btol => ::Real: btol stopping tolerance.

:ctol => ::Real: ctol stopping tolerance.

:anorm => ::Real: anorm.

:rnorm => ::Real: rnorm.

:cnorm => ::Real: cnorm.

:resnom => ::Vector: residual norm at each iteration.

source

IterativeSolvers.lsmr! — Function.

lsmr!(x, A, b)

Overwrite x.

Minimize ||Ax-b||^2 + λ^2 ||x||^2 for A*x=b.

The method is based on the Golub-Kahan bidiagonalization process. It is algebraically equivalent to applying MINRES to the normal equation (ATA+λ2I)x=ATb, but has better numerical properties, especially if A is ill-conditioned.

If log is set to true is given, method will output a tuple x, ch. Where ch is a ConvergenceHistory object. Otherwise it will only return x.

The plot attribute can only be used when log is set version.

Arguments

x: initial guess, overwrite final estimation.

A: linear operator.

b: right hand side.

Keywords

λ::Number = 0: lambda.

atol::Number = 1e-6, btol::Number = 1e-6: stopping tolerances. If both are 1.0e-9 (say), the final residual norm should be accurate to about 9 digits. (The final x will usually have fewer correct digits, depending on cond(A) and the size of damp).

conlim::Number = 1e8: stopping tolerance. lsmr terminates if an estimate of cond(A) exceeds conlim. For compatible systems Ax = b, conlim could be as large as 1.0e+12 (say). For least-squares problems, conlim should be less than 1.0e+8. Maximum precision can be obtained by setting atol = btol = conlim = zero, but the number of iterations may then be excessive.

maxiter::Integer = min(20,length(b)): maximum number of iterations.

verbose::Bool = false: print method information.

log::Bool = false: output an extra element of type ConvergenceHistory containing extra information of the method execution.

plot::Bool = false: plot data. (Only when log is set)

Output

if log is false

x: approximated solution.

if log is true

x: approximated solution.

ch: convergence history.

ConvergenceHistory keys

:atol => ::Real: atol stopping tolerance.

:btol => ::Real: btol stopping tolerance.

:ctol => ::Real: ctol stopping tolerance.

:anorm => ::Real: anorm.

:rnorm => ::Real: rnorm.

:cnorm => ::Real: cnorm.

:resnom => ::Vector: residual norm at each iteration.

source

References

Adapted from: http://web.stanford.edu/group/SOL/software/lsmr/

LSQR

IterativeSolvers.lsqr — Function.

lsqr(A, b)

LSQR solves Ax = b or min ||b - Ax||^2 if damp = 0, or min ||(b) - ( A )x|| otherwise. ||(0) (damp*I) ||^2.

The method is based on the Golub-Kahan bidiagonalization process. It is algebraically equivalent to applying CG to the normal equation (ATA+λ2I)x=ATb, but has better numerical properties, especially if A is ill-conditioned.

If log is set to true is given, method will output a tuple x, ch. Where ch is a ConvergenceHistory object. Otherwise it will only return x.

The plot attribute can only be used when log is set version.

Arguments

A: linear operator.

b: right hand side.

Keywords

λ::Number = 0: lambda.

atol::Number = 1e-6, btol::Number = 1e-6: stopping tolerances. If both are 1.0e-9 (say), the final residual norm should be accurate to about 9 digits. (The final x will usually have fewer correct digits, depending on cond(A) and the size of damp).

conlim::Number = 1e8: stopping tolerance. lsmr terminates if an estimate of cond(A) exceeds conlim. For compatible systems Ax = b, conlim could be as large as 1.0e+12 (say). For least-squares problems, conlim should be less than 1.0e+8. Maximum precision can be obtained by setting atol = btol = conlim = zero, but the number of iterations may then be excessive.

maxiter::Integer = min(20,length(b)): maximum number of iterations.

verbose::Bool = false: print method information.

log::Bool = false: output an extra element of type ConvergenceHistory containing extra information of the method execution.

plot::Bool = false: plot data. (Only when log is set)

Output

if log is false

x: approximated solution.

if log is true

x: approximated solution.

ch: convergence history.

ConvergenceHistory keys

:atol => ::Real: atol stopping tolerance.

:btol => ::Real: btol stopping tolerance.

:ctol => ::Real: ctol stopping tolerance.

:anorm => ::Real: anorm.

:rnorm => ::Real: rnorm.

:cnorm => ::Real: cnorm.

:resnom => ::Vector: residual norm at each iteration.

source

IterativeSolvers.lsqr! — Function.

lsqr!(x, A, b)

Overwrite x.

LSQR solves Ax = b or min ||b - Ax||^2 if damp = 0, or min ||(b) - ( A )x|| otherwise. ||(0) (damp*I) ||^2.

The method is based on the Golub-Kahan bidiagonalization process. It is algebraically equivalent to applying CG to the normal equation (ATA+λ2I)x=ATb, but has better numerical properties, especially if A is ill-conditioned.

If log is set to true is given, method will output a tuple x, ch. Where ch is a ConvergenceHistory object. Otherwise it will only return x.

The plot attribute can only be used when log is set version.

Arguments

x: initial guess, overwrite final estimation.

A: linear operator.

b: right hand side.

Keywords

λ::Number = 0: lambda.

atol::Number = 1e-6, btol::Number = 1e-6: stopping tolerances. If both are 1.0e-9 (say), the final residual norm should be accurate to about 9 digits. (The final x will usually have fewer correct digits, depending on cond(A) and the size of damp).

conlim::Number = 1e8: stopping tolerance. lsmr terminates if an estimate of cond(A) exceeds conlim. For compatible systems Ax = b, conlim could be as large as 1.0e+12 (say). For least-squares problems, conlim should be less than 1.0e+8. Maximum precision can be obtained by setting atol = btol = conlim = zero, but the number of iterations may then be excessive.

maxiter::Integer = min(20,length(b)): maximum number of iterations.

verbose::Bool = false: print method information.

log::Bool = false: output an extra element of type ConvergenceHistory containing extra information of the method execution.

plot::Bool = false: plot data. (Only when log is set)

Output

if log is false

x: approximated solution.

if log is true

x: approximated solution.

ch: convergence history.

ConvergenceHistory keys

:atol => ::Real: atol stopping tolerance.

:btol => ::Real: btol stopping tolerance.

:ctol => ::Real: ctol stopping tolerance.

:anorm => ::Real: anorm.

:rnorm => ::Real: rnorm.

:cnorm => ::Real: cnorm.

:resnom => ::Vector: residual norm at each iteration.

source

References

Adapted from: http://web.stanford.edu/group/SOL/software/lsqr/

1. C. C. Paige and M. A. Saunders (1982a).
    LSQR: An algorithm for sparse linear equations and sparse least squares,
    ACM TOMS 8(1), 43-71.

2. C. C. Paige and M. A. Saunders (1982b).
    Algorithm 583.  LSQR: Sparse linear equations and least squares problems,
    ACM TOMS 8(2), 195-209.

3. M. A. Saunders (1995).  Solution of sparse rectangular systems using
    LSQR and CRAIG, BIT 35, 588-604.

Conjugate gradients

IterativeSolvers.cg — Function.

cg(A, b)

Solve A*x=b with the conjugate gradients method.

If log is set to true is given, method will output a tuple x, ch. Where ch is a ConvergenceHistory object. Otherwise it will only return x. The plot attribute can only be used when log is set version.

Arguments

A: linear operator.

b: right hand side.

Keywords

Pl = 1: left preconditioner of the method.

tol::Real = size(A,2)*eps(): stopping tolerance.

maxiter::Integer = size(A,2): maximum number of iterations.

verbose::Bool = false: print method information.

log::Bool = false: output an extra element of type ConvergenceHistory containing extra information of the method execution.

plot::Bool = false: plot data. (Only when log is set)

Output

if log is false

x: approximated solution.

if log is true

x: approximated solution.

ch: convergence history.

ConvergenceHistory keys

:tol => ::Real: stopping tolerance.

:resnom => ::Vector: residual norm at each iteration.

source

IterativeSolvers.cg! — Function.

cg!(x, A, b)

Overwrite x.

Solve A*x=b with the conjugate gradients method.

If log is set to true is given, method will output a tuple x, ch. Where ch is a ConvergenceHistory object. Otherwise it will only return x. The plot attribute can only be used when log is set version.

Arguments

x: initial guess, overwrite final estimation.

A: linear operator.

b: right hand side.

Keywords

Pl = 1: left preconditioner of the method.

tol::Real = size(A,2)*eps(): stopping tolerance.

maxiter::Integer = size(A,2): maximum number of iterations.

verbose::Bool = false: print method information.

log::Bool = false: output an extra element of type ConvergenceHistory containing extra information of the method execution.

plot::Bool = false: plot data. (Only when log is set)

Output

if log is false

x: approximated solution.

if log is true

x: approximated solution.

ch: convergence history.

ConvergenceHistory keys

:tol => ::Real: stopping tolerance.

:resnom => ::Vector: residual norm at each iteration.

source

Chebyshev iteration

IterativeSolvers.chebyshev — Function.

chebyshev!(A, b, λmin, λmax)

Solve A*x=b using the chebyshev method.

If log is set to true is given, method will output a tuple x, ch. Where ch is a ConvergenceHistory object. Otherwise it will only return x.

The plot attribute can only be used when log is set version.

Arguments

A: linear operator.

b: right hand side.

Keywords

Pr = 1: right preconditioner of the method.

tol::Real = sqrt(eps()): stopping tolerance.

maxiter::Integer = size(A,2)^3: maximum number of iterations.

verbose::Bool = false: print method information.

log::Bool = false: output an extra element of type ConvergenceHistory containing extra information of the method execution.

plot::Bool = false: plot data. (Only with Master version)

Output

if log is false

x: approximated solution.

if log is true

x: approximated solution.

ch: convergence history.

ConvergenceHistory keys

:tol => ::Real: stopping tolerance.

:resnom => ::Vector: residual norm at each iteration.

source

IterativeSolvers.chebyshev! — Function.

chebyshev!(x, A, b, λmin, λmax)

Overwrite x.

Solve A*x=b using the chebyshev method.

If log is set to true is given, method will output a tuple x, ch. Where ch is a ConvergenceHistory object. Otherwise it will only return x.

The plot attribute can only be used when log is set version.

Arguments

x: initial guess, overwrite final estimation.

A: linear operator.

b: right hand side.

Keywords

Pr = 1: right preconditioner of the method.

tol::Real = sqrt(eps()): stopping tolerance.

maxiter::Integer = size(A,2)^3: maximum number of iterations.

verbose::Bool = false: print method information.

log::Bool = false: output an extra element of type ConvergenceHistory containing extra information of the method execution.

plot::Bool = false: plot data. (Only with Master version)

Output

if log is false

x: approximated solution.

if log is true

x: approximated solution.

ch: convergence history.

ConvergenceHistory keys

:tol => ::Real: stopping tolerance.

:resnom => ::Vector: residual norm at each iteration.

source

Generalized minimal residual method (with restarts)

IterativeSolvers.gmres — Function.

gmres(A, b)

Solve A*x=b using the generalized minimal residual method with restarts.

If log is set to true is given, method will output a tuple x, ch. Where ch is a ConvergenceHistory object. Otherwise it will only return x.

The plot attribute can only be used when log is set version.

Arguments

A: linear operator.

b: right hand side.

Keywords

Pl = 1: left preconditioner of the method.

Pr = 1: left preconditioner of the method.

tol::Real = sqrt(eps()): stopping tolerance.

restart::Integer = min(20,length(b)): maximum number of iterations per restart.

maxiter::Integer = min(20,length(b)): maximum number of iterations.

verbose::Bool = false: print method information.

log::Bool = false: output an extra element of type ConvergenceHistory containing extra information of the method execution.

plot::Bool = false: plot data. (Only when log is set)

Output

if log is false

x: approximated solution.

if log is true

x: approximated solution.

ch: convergence history.

ConvergenceHistory keys

:tol => ::Real: stopping tolerance. :resnom => ::Vector: residual norm at each iteration.

References

http://www.netlib.org/templates/templates.pdf 2.3.4 Generalized Minimal Residual (GMRES)

http://www.netlib.org/lapack/lawnspdf/lawn148.pdf Givens rotation based on Algorithm 1

source

IterativeSolvers.gmres! — Function.

gmres!(x, A, b)

Overwrite x.

Solve A*x=b using the generalized minimal residual method with restarts.

If log is set to true is given, method will output a tuple x, ch. Where ch is a ConvergenceHistory object. Otherwise it will only return x.

The plot attribute can only be used when log is set version.

Arguments

x: initial guess, overwrite final estimation.

A: linear operator.

b: right hand side.

Keywords

Pl = 1: left preconditioner of the method.

Pr = 1: left preconditioner of the method.

tol::Real = sqrt(eps()): stopping tolerance.

restart::Integer = min(20,length(b)): maximum number of iterations per restart.

maxiter::Integer = min(20,length(b)): maximum number of iterations.

verbose::Bool = false: print method information.

log::Bool = false: output an extra element of type ConvergenceHistory containing extra information of the method execution.

plot::Bool = false: plot data. (Only when log is set)

Output

if log is false

x: approximated solution.

if log is true

x: approximated solution.

ch: convergence history.

ConvergenceHistory keys

:tol => ::Real: stopping tolerance. :resnom => ::Vector: residual norm at each iteration.

References

http://www.netlib.org/templates/templates.pdf 2.3.4 Generalized Minimal Residual (GMRES)

http://www.netlib.org/lapack/lawnspdf/lawn148.pdf Givens rotation based on Algorithm 1

source

Eigen Solvers

Power iteration

IterativeSolvers.powm — Function.

powm(A)

Find biggest eigenvalue of A and its associated eigenvector using the power method.

If log is set to true is given, method will output a tuple eig, v, ch. Where ch is a ConvergenceHistory object. Otherwise it will only return eig, v.

The plot attribute can only be used when log is set version.

Arguments

K::KrylovSubspace: krylov subspace.

A: linear operator.

Keywords

x = random unit vector: initial eigenvector guess.

tol::Real = eps()*size(A,2)^3: stopping tolerance.

maxiter::Integer = size(A,2): maximum number of iterations.

verbose::Bool = false: verbose flag.

log::Bool = false: output an extra element of type ConvergenceHistory containing extra information of the method execution.

plot::Bool = false: plot data. (Only when log is set)

Output

if log is false

eig::Real: eigen value

v::Vector: eigen vector

if log is true

eig::Real: eigen value

v::Vector: eigen vector

ch: convergence history.

ConvergenceHistory keys

:tol => ::Real: stopping tolerance.

:resnom => ::Vector: residual norm at each iteration.

source

Inverse power iteration

IterativeSolvers.invpowm — Function.

invpowm(A)

Find closest eigenvalue of A to shift and its associated eigenvector using the inverse power iteration method.

If log is set to true is given, method will output a tuple eig, v, ch. Where ch is a ConvergenceHistory object. Otherwise it will only return eig, v.

The plot attribute can only be used when log is set version.

Arguments

K::KrylovSubspace: krylov subspace.

A: linear operator.

Keywords

shift::Number=0: shift to be applied to matrix A.

x = random unit vector: initial eigenvector guess.

tol::Real = eps()*size(A,2)^3: stopping tolerance.

maxiter::Integer = size(A,2): maximum number of iterations.

verbose::Bool = false: verbose flag.

log::Bool = false: output an extra element of type ConvergenceHistory containing extra information of the method execution.

plot::Bool = false: plot data. (Only when log is set)

Output

if log is false

eig::Real: eigen value

v::Vector: eigen vector

if log is true

eig::Real: eigen value

v::Vector: eigen vector

ch: convergence history.

ConvergenceHistory keys

:tol => ::Real: stopping tolerance.

:resnom => ::Vector: residual norm at each iteration.

source

Simple Lanczos

IterativeSolvers.eiglancz — Function.

eiglancz(A)

Find the most useful eigenvalues using the lanczos method.

If log is set to true is given, method will output a tuple eigs, ch. Where ch is a ConvergenceHistory object. Otherwise it will only return eigs.

The plot attribute can only be used when log is set version.

Arguments

A: linear operator.

Keywords

neigs::Int = size(A,1): number of eigen values.

tol::Real = size(A,1)^3*eps(): stopping tolerance.

maxiter::Integer=size(A,1): maximum number of iterations.

verbose::Bool = false: verbose flag.

log::Bool = false: output an extra element of type ConvergenceHistory containing extra information of the method execution.

plot::Bool = false: plot data. (Only when log is set)

Output

if log is false

eigs::Vector: eigen values.

if log is true

eigs::Vector: eigen values.

ch: convergence history.

ConvergenceHistory keys

:tol => ::Real: stopping tolerance.

:resnom => ::Vector: residual norm at each iteration.

source

Golub-Kahan-Lanczos

IterativeSolvers.svdl — Function.

svdl(A)

Compute some singular values (and optionally vectors) using Golub-Kahan-Lanczos bidiagonalization cite{Golub1965} with thick restarting cite{Wu2000}.

If log is set to true is given, method will output a tuple X, L, ch. Where ch is a ConvergenceHistory object. Otherwise it will only return X, L.

The plot attribute can only be used when log is set version.

Arguments

A : The matrix or matrix-like object whose singular values are desired.

Keywords

nsv::Int = 6: number of singular values requested.

v0 = random unit vector: starting guess vector in the domain of A. The length of q should be the number of columns in A.

k::Int = 2nsv: maximum number of Lanczos vectors to compute before restarting.

j::Int = nsv: number of vectors to keep at the end of the restart. We don't recommend j < nsv.

maxiter::Int = minimum(size(A)): maximum number of iterations to run.

verbose::Bool = false: print information at each iteration.

tol::Real = √eps(): maximum absolute error in each desired singular value.

reltol::Real=√eps(): maximum error in each desired singular value relative to the estimated norm of the input matrix.

method::Symbol=:ritz: restarting algorithm to use. Valid choices are:

:ritz: Thick restart with Ritz values [Wu2000].

:harmonic: Restart with harmonic Ritz values [Baglama2005].

vecs::Symbol = :none: singular vectors to return.

:both: Both left and right singular vectors are returned.

:left: Only the left singular vectors are returned.

:right: Only the right singular vectors are returned.

:none: No singular vectors are returned.

dolock::Bool=false: If true, locks converged Ritz values, removing them from the Krylov subspace being searched in the next macroiteration.

log::Bool = false: output an extra element of type ConvergenceHistory containing extra information of the method execution.

plot::Bool = false: plot data. (Only when log is set)

Output

if log is false

Σ: list of the desired singular values if vecs == :none (the default), otherwise returns an SVD object with the desired singular vectors filled in.

L: computed partial factorizations of A.

if log is true

Σ: list of the desired singular values if vecs == :none (the default), otherwise returns an SVD object with the desired singular vectors filled in.

L: computed partial factorizations of A.

ch::ConvergenceHistory: convergence history.

ConvergenceHistory keys

:betas => betas: The history of the computed betas.

:Bs => Bs: The history of the computed projected matrices.

:ritz => ritzvalhist: Ritz values computed at each iteration.

:conv => convhist: Convergence data.

source

Implementation notes

The implementation of thick restarting follows closely that of SLEPc as described in [Hernandez2008]. Thick restarting can be turned off by setting k = maxiter, but most of the time this is not desirable.

The singular vectors are computed directly by forming the Ritz vectors from the product of the Lanczos vectors L.P/L.Q and the singular vectors of L.B. Additional accuracy in the singular triples can be obtained using inverse iteration.

References

@article{Golub1965,
    author = {Golub, G. and Kahan, W.},
    doi = {10.1137/0702016},
    journal = {Journal of the Society for Industrial and Applied Mathematics
        Series B Numerical Analysis},
    volume = 2,
    number = 2,
    pages = {205--224},
    title = {Calculating the Singular Values and Pseudo-Inverse of a Matrix},
    year = 1965
}

@article{Wu2000,
    author = {Wu, Kesheng and Simon, Horst},
    journal = {SIAM Journal on Matrix Analysis and Applications},
    number = 2,
    pages = {602--616},
    title = {Thick-Restart {L}anczos Method for Large Symmetric Eigenvalue Problems},
    volume = 22,
    year = 2000
}

@article{Baglama2005,
    author = {Baglama, James and Reichel, Lothar},
    doi = {10.1137/04060593X},
    journal = {SIAM Journal on Scientific Computing},
    number = 1,
    pages = {19--42},
    title = {Augmented Implicitly Restarted {L}anczos Bidiagonalization Methods},
    volume = 27,
    year = 2005
}

@article{Hernandez2008,
    author = {Hern\'{a}ndez, Vicente and Rom\'{a}n, Jos\'{e} E and Tom\'{a}s,
    Andr\'{e}s},
    journal = {Electronic Transactions on Numerical Analysis},
    pages = {68--85},
    title = {A Robust and Efficient Parallel {SVD} Solver based on Restarted
        {L}anczos Bidiagonalization},
    url = {http://etna.mcs.kent.edu/volumes/2001-2010/vol31/abstract.php?vol=31\&pages=68-85},
    volume = 31,
    year = 2008
}

Randomized

Condition number estimate

IterativeSolvers.rcond — Function.

rcond(A, iters=1)

Estimate matrix condition number randomly.

Arguments

A: matrix whose condition number to estimate. Must be square and support premultiply (A*⋅) and solve (A⋅).

iters::Int = 1: number of power iterations to run.

Keywords

p::Real = 0.05: probability that estimate fails to hold as an upper bound.

Output

Interval (x, y) which contains κ(A) with probability 1 - p.

Implementation note

cite{Dixon1983} originally describes this as a computation that can be done by computing the necessary number of power iterations given p and the desired accuracy parameter θ=y/x. However, these bounds were only derived under the assumptions of exact arithmetic. Empirically, iters≥4 has been seen to result in incorrect results in that the computed interval does not contain the true condition number. This implemention therefore makes iters an explicitly user-controllable parameter from which to infer the accuracy parameter and hence the interval containing κ(A).

References

cite[Theorem 2]{Dixon1983}

@article{Dixon1983,
    author = {Dixon, John D},
    doi = {10.1137/0720053},
    journal = {SIAM Journal on Numerical Analysis},
    number = {4},
    pages = {812--814},
    title = {Estimating Extremal Eigenvalues and Condition Numbers of
	Matrices},
    volume = {20},
    year = {1983}
}

source

Extremal eigenvalue estimates

IterativeSolvers.reigmin — Function.

reigmin(A, iters=1)

Estimate minimal eigenvalue randomly.

Arguments

A: Matrix whose maximal eigenvalue to estimate. Must be square and support premultiply (A*⋅).

iters::Int=1: Number of power iterations to run. (Recommended: iters ≤ 3)

Keywords

p::Real=0.05: Probability that estimate fails to hold as an upper bound.

Output

Interval (x, y) which contains the maximal eigenvalue of A with probability 1 - p.

References

cite[Corollary of Theorem 1]{Dixon1983}.

source

IterativeSolvers.reigmax — Function.

reigmax(A, iters=1)

Estimate maximal eigenvalue randomly.

Arguments

A: Matrix whose maximal eigenvalue to estimate. Must be square and support premultiply (A*⋅).

iters::Int=1: Number of power iterations to run. (Recommended: iters ≤ 3)

Keywords

p::Real=0.05: Probability that estimate fails to hold as an upper bound.

Output

Interval (x, y) which contains the maximal eigenvalue of A with probability 1 - p.

References

cite[Corollary of Theorem 1]{Dixon1983}.

source

Norm estimate

IterativeSolvers.rnorm — Function.

rnorm(A, mvps)

Compute a probabilistic upper bound on the norm of a matrix A. ‖A‖ ≤ α √(2/π) maxᵢ ‖Aωᵢ‖ with probability p=α^(-mvps).

Arguments

A: matrix whose norm to estimate.

mvps::Int: number of matrix-vector products to compute.

Keywords

p::Real=0.05: probability of upper bound failing.

Output

Estimate of ‖A‖.

See also

see rnorms for a different estimator that uses premultiplying by both A and A'.

References

cite[Lemma 4.1]{Halko2011}

source

IterativeSolvers.rnorms — Function.

rnorms(A, iters=1)

Estimate matrix norm randomly using A'A.

Compute a probabilistic upper bound on the norm of a matrix A.

ρ = √(‖(A'A)ʲω‖/‖(A'A)ʲ⁻¹ω‖)

which is an estimate of the spectral norm of A produced by iters steps of the power method starting with normalized ω, is a lower bound on the true norm by a factor

ρ ≤ α ‖A‖

with probability greater than 1 - p, where p = 4sqrt(n/(iters-1)) α^(-2iters).

Arguments

A: matrix whose norm to estimate.

iters::Int = 1: mumber of power iterations to perform.

Keywords

p::Real = 0.05: probability of upper bound failing.

At = A': Transpose of A.

Output

Estimate of ‖A‖.

See also

see rnorm for a different estimator that does not require premultiplying by A'

References

Appendix of cite{Liberty2007}.

@article{Liberty2007,
    authors = {Edo Liberty and Franco Woolfe and Per-Gunnar Martinsson
    and Vladimir Rokhlin and Mark Tygert},
    title = {Randomized algorithms for the low-rank approximation of matrices},
    journal = {Proceedings of the National Academy of Sciences},
    volume = {104},
    issue = {51},
    year = {2007},
    pages = {20167--20172},
    doi  = {10.1073/pnas.0709640104}
}

source

Randomized singular value decomposition

IterativeSolvers.reig — Function.

reig(A, l)

Compute the spectral (Eigen) decomposition of A using a randomized algorithm.

Arguments

A: input matrix.

l::Int: number of eigenpairs to find.

Output

::Base.LinAlg.Eigen: eigen decomposition.

Implementation note

This is a wrapper around eigfact_onepass() which uses the randomized samples found using srft(l).

References

@article{Halko2011,
    author = {Halko, N and Martinsson, P G and Tropp, J A},
    doi = {10.1137/090771806},
    journal = {SIAM Review},
    month = jan,
    number = {2},
    pages = {217--288},
    title = {Finding Structure with Randomness: Probabilistic Algorithms for Constructing Approximate Matrix Decompositions},
    volume = {53},
    year = {2011}
}

source

IterativeSolvers.rsvdfact — Function.

rsvdfact(A, n, p=0)

Compute partial singular value decomposition of A using a randomized algorithm.

Arguments

A: input matrix.

n::Int: number of singular value/vector pairs to find.

p::Int=0: number of extra vectors to include in computation.

Output

::SVD: singular value decomposition.

Warning

This variant of the randomized singular value decomposition is the most commonly found implementation but is not recommended for accurate computations, as it often has trouble finding the n largest singular pairs, but rather finds n large singular pairs which may not necessarily be the largest.

Implementation note

This function calls rrange, which uses naive randomized rangefinding to compute a basis for a subspace of dimension n (Algorithm 4.1 of cite{Halko2011}), followed by svdfact_restricted(), which computes the exact SVD factorization on the restriction of A to this randomly selected subspace (Algorithm 5.1 of cite{Halko2011}).

Alternatively, you can mix and match your own randomized algorithm using any of the randomized range finding algorithms to find a suitable subspace and feeding the result to one of the routines that computes the SVD restricted to that subspace.

References

@article{Halko2011,
    author = {Halko, N and Martinsson, P G and Tropp, J A},
    doi = {10.1137/090771806},
    journal = {SIAM Review},
    month = jan,
    number = {2},
    pages = {217--288},
    title = {Finding Structure with Randomness: Probabilistic Algorithms for Constructing Approximate Matrix Decompositions},
    volume = {53},
    year = {2011}
}

source

IterativeSolvers.rsvd_fnkz — Function.

rsvd_fnkz(A, k)

Compute the randomized SVD by iterative refinement from randomly selected columns/rows.

Arguments

A: matrix whose SVD is desired.

k::Int: desired rank of approximation (k ≤ min(m, n)).

Keywords

l::Int = k: number of columns/rows to sample at each iteration (1 ≤ l ≤ k).

N::Int = minimum(size(A)): maximum number of iterations.

ϵ::Real = prod(size(A))*eps(): relative threshold for convergence, as measured by growth of the spectral norm.

method::Symbol = :eig: problem to solve.

:eig: eigenproblem.

:svd: singular problem.

verbose::Bool = false: print convergence information at each iteration.

Output

SVD object of rank ≤ k.

References

@inproceedings{,
    author={Friedland, S. and Niknejad, A. and Kaveh, Mostafa and Zare, H.},
    booktitle={System of Systems Engineering, 2006 IEEE/SMC International Conference on},
    title={Fast Monte-Carlo low rank approximations for matrices},
    year={2006},
    month={April},
    pages={218--223},
    doi={10.1109/SYSOSE.2006.1652299}
}

source