Stationary methods

Stationary methods

Stationary methods are typically used as smoothers in multigrid methods, where only very few iterations are applied to get rid of high-frequency components in the error. The implementations of stationary methods have this goal in mind, which means there is no other stopping criterion besides the maximum number of iterations.

CSC versus CSR

Julia stores matrices column-major. In order to avoid cache misses, the implementations of our stationary methods traverse the matrices column-major. This deviates from classical textbook implementations. Also the (S)SOR methods cannot be computed fully in-place, but require a temporary vector.

When it comes to SparseMatrixCSC, we precompute an integer array of the indices of the diagonal as well to avoid expensive searches in each iteration.

Jacobi

IterativeSolvers.jacobi — Function.

jacobi(A, b)

Solve A*x=b with the Jacobi method.

ch is a ConvergenceHistory object. Otherwise it will only return x. If log is set to true is given, method will output a tuple x, ch.

Arguments

A: linear operator.

Keywords

tol::Real = size(A,2)^3*eps(): stopping tolerance.

maxiter::Integer = size(A,2)^2: maximum number of iterations.

verbose::Bool = false: verbose flag.

log::Bool = false: output an extra element of type ConvergenceHistory containing extra information of the method execution.

Output

if log is false

x: approximated solution.

if log is true

x: approximated solution.

ch: convergence history.

ConvergenceHistory keys

:tol => ::Real: stopping tolerance.

:resnom => ::Vector: residual norm at each iteration.

source

IterativeSolvers.jacobi! — Function.

jacobi!(x, A, b)

Overwrite x.

Solve A*x=b with the Jacobi method.

ch is a ConvergenceHistory object. Otherwise it will only return x. If log is set to true is given, method will output a tuple x, ch.

Arguments

x: initial guess, overwrite final estimation.

A: linear operator.

Keywords

tol::Real = size(A,2)^3*eps(): stopping tolerance.

maxiter::Integer = size(A,2)^2: maximum number of iterations.

verbose::Bool = false: verbose flag.

log::Bool = false: output an extra element of type ConvergenceHistory containing extra information of the method execution.

Output

if log is false

x: approximated solution.

if log is true

x: approximated solution.

ch: convergence history.

ConvergenceHistory keys

:tol => ::Real: stopping tolerance.

:resnom => ::Vector: residual norm at each iteration.

source

Gauss-Seidel

IterativeSolvers.gauss_seidel — Function.

gauss_seidel(A, b)

Solve A*x=b with the Gauss-Seidel method.

ch is a ConvergenceHistory object. Otherwise it will only return x. If log is set to true is given, method will output a tuple x, ch.

Arguments

A: linear operator.

Keywords

tol::Real = size(A,2)^3*eps(): stopping tolerance.

maxiter::Integer = size(A,2)^2: maximum number of iterations.

verbose::Bool = false: verbose flag.

log::Bool = false: output an extra element of type ConvergenceHistory containing extra information of the method execution.

Output

if log is false

x: approximated solution.

if log is true

x: approximated solution.

ch: convergence history.

ConvergenceHistory keys

:tol => ::Real: stopping tolerance.

:resnom => ::Vector: residual norm at each iteration.

source

IterativeSolvers.gauss_seidel! — Function.

gauss_seidel!(x, A, b)

Overwrite x.

Solve A*x=b with the Gauss-Seidel method.

ch is a ConvergenceHistory object. Otherwise it will only return x. If log is set to true is given, method will output a tuple x, ch.

Arguments

x: initial guess, overwrite final estimation.

A: linear operator.

Keywords

tol::Real = size(A,2)^3*eps(): stopping tolerance.

maxiter::Integer = size(A,2)^2: maximum number of iterations.

verbose::Bool = false: verbose flag.

log::Bool = false: output an extra element of type ConvergenceHistory containing extra information of the method execution.

Output

if log is false

x: approximated solution.

if log is true

x: approximated solution.

ch: convergence history.

ConvergenceHistory keys

:tol => ::Real: stopping tolerance.

:resnom => ::Vector: residual norm at each iteration.

source

Successive over-relaxation (SOR)

IterativeSolvers.sor — Function.

sor(A, b, ω)

Solve A*x=b with the successive overrelaxation method.

ch is a ConvergenceHistory object. Otherwise it will only return x. If log is set to true is given, method will output a tuple x, ch.

Arguments

shift::Number=0: shift to be applied to matrix A.

A: linear operator.

Keywords

tol::Real = size(A,2)^3*eps(): stopping tolerance.

maxiter::Integer = size(A,2)^2: maximum number of iterations.

verbose::Bool = false: verbose flag.

log::Bool = false: output an extra element of type ConvergenceHistory containing extra information of the method execution.

Output

if log is false

x: approximated solution.

if log is true

x: approximated solution.

ch: convergence history.

ConvergenceHistory keys

:tol => ::Real: stopping tolerance.

:resnom => ::Vector: residual norm at each iteration.

source

IterativeSolvers.sor! — Function.

sor!(x, A, b, ω)

Overwrite x.

Solve A*x=b with the successive overrelaxation method.

ch is a ConvergenceHistory object. Otherwise it will only return x. If log is set to true is given, method will output a tuple x, ch.

Arguments

x: initial guess, overwrite final estimation.

shift::Number=0: shift to be applied to matrix A.

A: linear operator.

Keywords

tol::Real = size(A,2)^3*eps(): stopping tolerance.

maxiter::Integer = size(A,2)^2: maximum number of iterations.

verbose::Bool = false: verbose flag.

log::Bool = false: output an extra element of type ConvergenceHistory containing extra information of the method execution.

Output

if log is false

x: approximated solution.

if log is true

x: approximated solution.

ch: convergence history.

ConvergenceHistory keys

:tol => ::Real: stopping tolerance.

:resnom => ::Vector: residual norm at each iteration.

source

Symmetric successive over-relaxation (SSOR)

IterativeSolvers.ssor — Function.

ssor(A, b, ω)

Solve A*x=b with the symmetric successive overrelaxation method.

ch is a ConvergenceHistory object. Otherwise it will only return x. If log is set to true is given, method will output a tuple x, ch.

Arguments

shift::Number=0: shift to be applied to matrix A.

A: linear operator.

Keywords

tol::Real = size(A,2)^3*eps(): stopping tolerance.

maxiter::Integer = size(A,2)^2: maximum number of iterations.

verbose::Bool = false: verbose flag.

log::Bool = false: output an extra element of type ConvergenceHistory containing extra information of the method execution.

Output

if log is false

x: approximated solution.

if log is true

x: approximated solution.

ch: convergence history.

ConvergenceHistory keys

:tol => ::Real: stopping tolerance.

:resnom => ::Vector: residual norm at each iteration.

source

IterativeSolvers.ssor! — Function.

ssor!(x, A, b, ω)

Overwrite x.

Solve A*x=b with the symmetric successive overrelaxation method.

ch is a ConvergenceHistory object. Otherwise it will only return x. If log is set to true is given, method will output a tuple x, ch.

Arguments

x: initial guess, overwrite final estimation.

shift::Number=0: shift to be applied to matrix A.

A: linear operator.

Keywords

tol::Real = size(A,2)^3*eps(): stopping tolerance.

maxiter::Integer = size(A,2)^2: maximum number of iterations.

verbose::Bool = false: verbose flag.

log::Bool = false: output an extra element of type ConvergenceHistory containing extra information of the method execution.

Output

if log is false

x: approximated solution.

if log is true

x: approximated solution.

ch: convergence history.

ConvergenceHistory keys

:tol => ::Real: stopping tolerance.

:resnom => ::Vector: residual norm at each iteration.

source